swc_config::merge

Trait Merge

pub trait Merge: Sized {
    // Required method
    fn merge(&mut self, other: Self);
}

Deriavable trait for overrding configurations.

Typically, correct implementation of this trait for a struct is calling merge for all fields, and #[derive(Merge)] will do it for you.

Required Methods

fn merge(&mut self, other: Self)

self has higher priority.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.